[REFACTOR] 서재 관심있어요 Application으로 로직 분리
